+++ packages/pygsti/objects/mapforwardsim.py
@@ -1033,8 +1033,9 @@ class MapForwardSimulator(ForwardSimulator):
             def calc_and_fill(rholabel, elabels, fIndsList, gIndsList, pslc1, pslc2, sumInto):
                 """ Compute and fill result quantities for given arguments """
                 #Fill cache info
-                chi2Cache = self.DM_compute_TDchi2_cache(rholabel, elabels, num_outcomes, evalSubTree, dataset_rows,
-                                                         minProbClipForWeighting, probClipInterval, mySubComm)
+                chi2Cache = replib.DM_compute_TDchi2_cache(self, rholabel, elabels, num_outcomes,
+                                                           evalSubTree, dataset_rows, minProbClipForWeighting,
+                                                           probClipInterval, mySubComm)
 
                 #use cached data to final values
                 ps = evalSubTree.final_view(chi2Cache, axis=0)  # ( nCircuits, len(elabels))
@@ -1050,75 +1051,6 @@ class MapForwardSimulator(ForwardSimulator):
                              mxToFill, [], 0, comm)
         #note: pass mxToFill, dim=(KS,), so gather mxToFill[felInds] (axis=0)
 
-    #TODO: move to slowreplib.py once we have a fast version
-    def DM_compute_TDchi2_cache(self, rholabel, elabels, num_outcomes, evalTree, dataset_rows,
-                                minProbClipForWeighting, probClipInterval, comm):
-
-        cacheSize = evalTree.cache_size()
-        rhoVec, EVecs = self._rhoEs_from_labels(rholabel, elabels)
-        ret = _np.zeros((len(evalTree), len(elabels)), 'd')  # zeros so we can just add contributions below
-
-        elabels_as_outcomes = [_model.spamTupleToOutcome((rholabel, e)) for e in elabels]
-        outcome_to_elabel_index = {outcome: i for i, outcome in enumerate(elabels_as_outcomes)}
-
-        assert(cacheSize == 0)  # so all elements have None as start and all as remainder
-        #if clipTo is not None:
-        #    _np.clip(mxToFill, clipTo[0], clipTo[1], out=mxToFill)  # in-place clip
-
-        #comm is currently ignored
-        #TODO: if evalTree is split, distribute among processors
-        for i in evalTree.get_evaluation_order():
-            iStart, remainder, iCache = evalTree[i]
-            assert(iStart is None), "Cannot use trees with max-cache-size > 0 when performing time-dependent calcs!"
-            datarow = dataset_rows[i]
-            nTotOutcomes = num_outcomes[i]
-
-            totalCnts = {}  # TODO defaultdict?
-            lastInds = {}; outcome_cnts = {}
-            for k, (t0, Nreps) in enumerate(zip(datarow.time, datarow.reps)):  # consolidate multiple outcomes that occur at same time? or sort?
-                if t0 in totalCnts:
-                    totalCnts[t0] += Nreps; outcome_cnts[t0] += 1
-                else:
-                    totalCnts[t0] = Nreps; outcome_cnts[t0] = 1
-                lastInds[t0] = k
-
-            cur_probtotal = 0; last_t = 0
-            for k, (t0, Nreps, outcome) in enumerate(zip(datarow.time, datarow.reps, datarow.outcomes)):  # consolidate multiple outcomes that occur at same time? or sort?
-                t = t0
-                rhoVec.set_time(t)
-                rho = rhoVec.torep('prep')
-                t += rholabel.time
-                
-                for gl in remainder:
-                    op = self.sos.get_operation(gl)
-                    op.set_time(t); t += gl.time  # time in gate label == gate duration?
-                    rho = op.torep().acton(rho)
-
-                j = outcome_to_elabel_index[outcome]
-                E = EVecs[j]; E.set_time(t)
-                p = E.torep('effect').probability(rho)  # outcome probability
-                cp = _np.clip(p, minProbClipForWeighting, 1 - minProbClipForWeighting)
-                N = totalCnts[t0]
-                f = Nreps / N
-                v = (p - f) * _np.sqrt(N / cp)
-
-                if t0 == last_t:
-                    cur_probtotal += p
-                else:
-                    last_t = t0
-                    cur_probtotal = p
-                
-                if lastInds[t0] == k and outcome_cnts[t0] < nTotOutcomes:  # cur_probtotal < 1.0?
-                    omitted_p = 1.0 - cur_probtotal
-                    omitted_cp = _np.clip(omitted_p, minProbClipForWeighting, 1 - minProbClipForWeighting)
-                    v = _np.sqrt(v**2 + N * omitted_p**2 / omitted_cp)
-                    # if this is the *last* outcome at this time then account for any omitted probability
-
-                ret[i, j] += v
-
-        return ret
-
-
     def bulk_fill_timedep_dchi2(self, mxToFill, evalTree, dsCircuitsToUse, num_total_outcomes, dataset,
                                 minProbClipForWeighting, probClipInterval, chi2MxToFill=None,
                                 comm=None, wrtFilter=None, wrtBlockSize=None,
@@ -1132,29 +1064,16 @@ class MapForwardSimulator(ForwardSimulator):
         None
         """
         def dchi2(rholabel, elabels, num_tot_outcomes, evalSubTree, dataset_rows, wrtSlice, fillComm):
-            return self.DM_compute_TDdchi2_cache(rholabel, elabels, num_tot_outcomes, evalSubTree, dataset_rows,
-                                                 minProbClipForWeighting, probClipInterval, wrtSlice, fillComm)
+            return replib.DM_compute_TDdchi2_cache(self, rholabel, elabels, num_tot_outcomes, evalSubTree, dataset_rows,
+                                                   minProbClipForWeighting, probClipInterval, wrtSlice, fillComm)
 
         def chi2(rholabel, elabels, num_tot_outcomes, evalSubTree, dataset_rows, fillComm):
-            return self.DM_compute_TDchi2_cache(rholabel, elabels, num_tot_outcomes, evalSubTree, dataset_rows,
-                                                minProbClipForWeighting, probClipInterval, fillComm)
+            return replib.DM_compute_TDchi2_cache(self, rholabel, elabels, num_tot_outcomes, evalSubTree, dataset_rows,
+                                                  minProbClipForWeighting, probClipInterval, fillComm)
 
         return self.bulk_fill_timedep_deriv(evalTree, dataset, dsCircuitsToUse, num_total_outcomes,
                                             mxToFill, dchi2, chi2MxToFill, chi2,
                                             comm, wrtFilter, wrtBlockSize, profiler, gatherMemLimit)
-
-
-    #TODO: move to slowreplib.py once we have a fast version
-    def DM_compute_TDdchi2_cache(self, rholabel, elabels, num_outcomes, evalTree, dataset_rows,
-                                 minProbClipForWeighting, probClipInterval, wrtSlice, comm):
-
-        def cachefn(rholabel, elabels, n_outcomes, evTree, dataset_rows, fillComm):
-            return self.DM_compute_TDchi2_cache(rholabel, elabels, n_outcomes, evTree, dataset_rows,
-                                                minProbClipForWeighting, probClipInterval, fillComm)
-
-        return self.DM_compute_timedep_dcache(rholabel, elabels, num_outcomes, evalTree, dataset_rows,
-                                              cachefn, wrtSlice, comm)
-
 
     def bulk_fill_timedep_loglpp(self, mxToFill, evalTree, dsCircuitsToUse, num_total_outcomes, dataset,
                                  minProbClip, radius, probClipInterval, comm=None):
@@ -1179,8 +1098,9 @@ class MapForwardSimulator(ForwardSimulator):
             def calc_and_fill(rholabel, elabels, fIndsList, gIndsList, pslc1, pslc2, sumInto):
                 """ Compute and fill result quantities for given arguments """
                 #Fill cache info
-                loglCache = self.DM_compute_TDloglpp_cache(rholabel, elabels, num_outcomes, evalSubTree, dataset_rows,
-                                                           minProbClip, radius, probClipInterval, mySubComm)
+                loglCache = replib.DM_compute_TDloglpp_cache(self, rholabel, elabels, num_outcomes,
+                                                             evalSubTree, dataset_rows, minProbClip,
+                                                             radius, probClipInterval, mySubComm)
 
                 #use cached data to final values
                 logl = evalSubTree.final_view(loglCache, axis=0)  # ( nCircuits, len(elabels))
@@ -1196,98 +1116,6 @@ class MapForwardSimulator(ForwardSimulator):
                              mxToFill, [], 0, comm)
         #note: pass mxToFill, dim=(KS,), so gather mxToFill[felInds] (axis=0)
 
-
-    #TODO: move to slowreplib.py once we have a fast version
-    def DM_compute_TDloglpp_cache(self, rholabel, elabels, num_outcomes, evalTree, dataset_rows,
-                                  minProbClip, radius, probClipInterval, comm):
-
-        min_p = minProbClip; a = radius
-        
-        cacheSize = evalTree.cache_size()
-        rhoVec, EVecs = self._rhoEs_from_labels(rholabel, elabels)
-        ret = _np.zeros((len(evalTree), len(elabels)), 'd')  # zeros so we can just add contributions below
-
-        elabels_as_outcomes = [_model.spamTupleToOutcome((rholabel, e)) for e in elabels]
-        outcome_to_elabel_index = {outcome: i for i, outcome in enumerate(elabels_as_outcomes)}
-
-        assert(cacheSize == 0)  # so all elements have None as start and all as remainder
-
-        #comm is currently ignored
-        #TODO: if evalTree is split, distribute among processors
-        for i in evalTree.get_evaluation_order():
-            iStart, remainder, iCache = evalTree[i]
-            assert(iStart is None), "Cannot use trees with max-cache-size > 0 when performing time-dependent calcs!"
-            datarow = dataset_rows[i]
-            nTotOutcomes = num_outcomes[i]
-
-            totalCnts = {}  # TODO defaultdict?
-            lastInds = {}; outcome_cnts = {}
-            for k, (t0, Nreps) in enumerate(zip(datarow.time, datarow.reps)):  # consolidate multiple outcomes that occur at same time? or sort?
-                if t0 in totalCnts:
-                    totalCnts[t0] += Nreps; outcome_cnts[t0] += 1
-                else:
-                    totalCnts[t0] = Nreps; outcome_cnts[t0] = 1
-                lastInds[t0] = k
-
-            cur_probtotal = 0; last_t = 0
-            for k, (t0, Nreps, outcome) in enumerate(zip(datarow.time, datarow.reps, datarow.outcomes)):  # consolidate multiple outcomes that occur at same time? or sort?
-                t = t0
-                rhoVec.set_time(t)
-                rho = rhoVec.torep('prep')
-                t += rholabel.time
-                
-                for gl in remainder:
-                    op = self.sos.get_operation(gl)
-                    op.set_time(t); t += gl.time  # time in gate label == gate duration?
-                    rho = op.torep().acton(rho)
-
-                j = outcome_to_elabel_index[outcome]
-                E = EVecs[j]; E.set_time(t)
-                p = E.torep('effect').probability(rho)  # outcome probability
-                N = totalCnts[t0]
-                f = Nreps / N
-                mNreps = -Nreps
-
-                #Compute logl term:
-                pos_p = max(p, min_p)
-                if Nreps != 0:
-                    freq_term = Nreps * (_np.log(f) - 1.0)
-                else:
-                    freq_term = 0.0
-                S = mNreps / min_p + N
-                S2 = -0.5 * mNreps / (min_p**2)
-                v = freq_term + mNreps * _np.log(pos_p) + N * pos_p  # dims K x M (K = nSpamLabels, M = nCircuits)
-
-                # remove small negative elements due to roundoff error (above expression *cannot* really be negative)
-                v = max(v, 0)
-                # quadratic extrapolation of logl at min_p for probabilities < min_p
-                if p < min_p:
-                    v = v + S * (p - min_p) + S2 * (p - min_p)**2
-
-                if mNreps == 0:
-                    if p >= a:
-                        v = N * p
-                    else:
-                        v = N * ((-1.0 / (3 * a**2)) * p**3 + p**2 / a + a / 3.0)
-                # special handling for f == 0 terms
-                # using quadratic rounding of function with minimum: max(0,(a-p)^2)/(2a) + p
-                
-                ret[i, j] += v
-
-                if t0 == last_t:
-                    cur_probtotal += p
-                else:
-                    last_t = t0
-                    cur_probtotal = p
-
-                if lastInds[t0] == k and outcome_cnts[t0] < nTotOutcomes:  # cur_probtotal < 1.0?
-                    omitted_p = 1.0 - cur_probtotal
-                    ret[i, j] += N * omitted_p if omitted_p >= a else \
-                        N * ((-1.0 / (3 * a**2)) * omitted_p**3 + omitted_p**2 / a + a / 3.0)
-                    # if this is the *last* outcome at this time then account for any omitted probability
-
-        return ret
-
     def bulk_fill_timedep_dloglpp(self, mxToFill, evalTree, dsCircuitsToUse, num_total_outcomes, dataset,
                                   minProbClip, radius, probClipInterval, loglMxToFill=None,
                                   comm=None, wrtFilter=None, wrtBlockSize=None,
@@ -1301,12 +1129,13 @@ class MapForwardSimulator(ForwardSimulator):
         None
         """
         def dloglpp(rholabel, elabels, num_tot_outcomes, evalSubTree, dataset_rows, wrtSlice, fillComm):
-            return self.DM_compute_TDdloglpp_cache(rholabel, elabels, num_tot_outcomes, evalSubTree, dataset_rows,
-                                                   minProbClip, radius, probClipInterval, wrtSlice, fillComm)
+            return replib.DM_compute_TDdloglpp_cache(self, rholabel, elabels, num_tot_outcomes, evalSubTree,
+                                                     dataset_rows, minProbClip, radius, probClipInterval,
+                                                     wrtSlice, fillComm)
 
         def loglpp(rholabel, elabels, num_tot_outcomes, evalSubTree, dataset_rows, fillComm):
-            return self.DM_compute_TDloglpp_cache(rholabel, elabels, num_tot_outcomes, evalSubTree, dataset_rows,
-                                                  minProbClip, radius, probClipInterval, fillComm)
+            return replib.DM_compute_TDloglpp_cache(self, rholabel, elabels, num_tot_outcomes, evalSubTree,
+                                                    dataset_rows, minProbClip, radius, probClipInterval, fillComm)
 
         return self.bulk_fill_timedep_deriv(evalTree, dataset, dsCircuitsToUse, num_total_outcomes,
                                             mxToFill, dloglpp, loglMxToFill, loglpp,
@@ -1431,67 +1260,3 @@ class MapForwardSimulator(ForwardSimulator):
         #profiler.add_time("bulk_fill_timedep_dchi2: total", tStart)
         #profiler.add_count("bulk_fill_timedep_dchi2 count")
         #profiler.mem_check("bulk_fill_timedep_dchi2: end")
-
-
-    #TODO: move to slowreplib.py once we have a fast version
-    def DM_compute_TDdloglpp_cache(self, rholabel, elabels, num_outcomes, evalTree, dataset_rows,
-                                   minProbClip, radius, probClipInterval, wrtSlice, comm):
-
-        def cachefn(rholabel, elabels, n_outcomes, evTree, dataset_rows, fillComm):
-            return self.DM_compute_TDloglpp_cache(rholabel, elabels, n_outcomes, evTree, dataset_rows,
-                                                  minProbClip, radius, probClipInterval, fillComm)
-        
-        return self.DM_compute_timedep_dcache(rholabel, elabels, num_outcomes, evalTree, dataset_rows,
-                                              cachefn, wrtSlice, comm)
-
-    #Generic function for base class
-    def DM_compute_timedep_dcache(self, rholabel, elabels, num_outcomes, evalTree, dataset_rows,
-                                  cachefn, wrtSlice, comm):
-
-        eps = 1e-7  # hardcoded?
-    
-        #Compute finite difference derivatives, one parameter at a time.
-        param_indices = range(self.Np) if (wrtSlice is None) else _slct.indices(wrtSlice)
-        nDerivCols = len(param_indices)  # *all*, not just locally computed ones
-    
-        rhoVec, EVecs = self._rhoEs_from_labels(rholabel, elabels)
-        cache = _np.empty((len(evalTree), len(elabels)), 'd')
-        dcache = _np.zeros((len(evalTree), len(elabels), nDerivCols), 'd')
-    
-        cacheSize = evalTree.cache_size()
-        assert(cacheSize == 0)
-    
-        cache = cachefn(rholabel, elabels, num_outcomes, evalTree, dataset_rows, comm)
-    
-        all_slices, my_slice, owners, subComm = \
-            _mpit.distribute_slice(slice(0, len(param_indices)), comm)
-    
-        my_param_indices = param_indices[my_slice]
-        st = my_slice.start  # beginning of where my_param_indices results
-        # get placed into dpr_cache
-    
-        #Get a map from global parameter indices to the desired
-        # final index within dpr_cache
-        iParamToFinal = {i: st + ii for ii, i in enumerate(my_param_indices)}
-    
-        orig_vec = self.to_vector().copy()
-        for i in range(self.Np):
-            #print("dprobs cache %d of %d" % (i,self.Np))
-            if i in iParamToFinal:
-                iFinal = iParamToFinal[i]
-                vec = orig_vec.copy(); vec[i] += eps
-                self.from_vector(vec)
-                dcache[:, :, iFinal] = (cachefn(rholabel, elabels, num_outcomes, evalTree, dataset_rows, subComm)
-                                        - cache) / eps
-        self.from_vector(orig_vec)
-    
-        #Now each processor has filled the relavant parts of dpr_cache,
-        # so gather together:
-        _mpit.gather_slices(all_slices, owners, dcache, [], axes=2, comm=comm)
-
-        #REMOVE
-        # DEBUG LINE USED FOR MONITORION N-QUBIT GST TESTS
-        #print("DEBUG TIME: dpr_cache(Np=%d, dim=%d, cachesize=%d, treesize=%d, napplies=%d) in %gs" %
-        #      (self.Np, self.dim, cacheSize, len(evalTree), evalTree.get_num_applies(), _time.time()-tStart)) #DEBUG
-    
-        return dcache
